Can now parse .epub files
Have modified the Book class to take in a File type and then parse it in the constructor and then return a Book class.
Have edited the Grid and List Tiles to take in the new object shape.
BookView has been changed to a FutureBuilder - to read the files in the EBookReader folder -> convert to Book objects -> return a list of books to the BookGrid and BookList widgets
TODO: Need to change this to a bloc and move it out of the BookView Widget